Introduce first-class `svg` module in `iced_native`
