+++ graphics/src/backend.rs
@@ -1,7 +1,7 @@
 //! Write a graphics backend.
 use iced_native::image;
+use iced_native::svg;
 use iced_native::text;
-use iced_native::widget::svg;
 use iced_native::{Font, Point, Size};
 
 /// The graphics backend of a [`Renderer`].
